test: remove the mockauthorizer
Remove the `MockAuthorizer` and instead initialize in the `MockServerModule`,
since it should be mocked in the tests
